Adjustments in Cliente entity

+++ SalesManagement/SalesManagement.Domain/Entities/Cliente.cs
@@ -4,7 +4,7 @@ namespace SalesManagement.Domain.Entities
 {
     public class Cliente
     {
-        public int ClienteId { get; set; }
+        public long ClienteId { get; set; }
 
         public string Nome { get; set; }
